license_portal/requirements.txt
Django==3.2
djangorestframework
mysqlclient
